refactor(arch): 単一のプロジェクトを Domain、Application、Infrastructure の3プロジェクトに分割
- ドメインモデル（Bukken等）のEF Core属性を排除し、Poco化（ScrapingContextのFluent APIへ移植）。
- ViewModels配下にあったDTOやStringExtensionをドメイン層へ移行して逆依存を排除。
