+++ .vscode/tasks.json
@@ -7,7 +7,7 @@
         "type": "process",
         "args": [
             "build",
-            "${workspaceFolder}/SuumoScraping.csproj",
+            "${workspaceFolder}/workspace.sln",
             "/property:GenerateFullPaths=true",
             "/consoleloggerparameters:NoSummary"
         ],
@@ -19,7 +19,7 @@
         "type": "process",
         "args": [
             "publish",
-            "${workspaceFolder}/SuumoScraping.csproj",
+            "${workspaceFolder}/workspace.sln",
             "/property:GenerateFullPaths=true",
             "/consoleloggerparameters:NoSummary"
         ],
