+++ Startup.cs
@@ -12,11 +12,11 @@ namespace SuumoScraping
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.Hosting;
     using Microsoft.Extensions.Logging;
+    using SuumoScraping.Application.UseCases;
     using SuumoScraping.Domain.Gateways;
     using SuumoScraping.Domain.Services;
     using SuumoScraping.Infrastructure.Persistence;
     using SuumoScraping.Infrastructure.Scraping;
-    using SuumoScraping.UseCases;
 
     public class Startup
     {
