+++ SuumoScraping.Application/UseCases/GetBukkenDetailsUseCase.cs
@@ -1,29 +1,27 @@
-namespace SuumoScraping.UseCases
+namespace SuumoScraping.Application.UseCases
 {
-    using System.Collections.Generic;
     using System.Threading;
     using System.Threading.Tasks;
     using SuumoScraping.Domain.Gateways;
-    using SuumoScraping.ViewModels;
+    using SuumoScraping.Domain.Models;
 
-    public class GetFilteredBukkensUseCase
+    public class GetBukkenDetailsUseCase
     {
         private readonly IScrapingContextFactory _scrapingContextFactory;
 
-        public GetFilteredBukkensUseCase(IScrapingContextFactory scrapingContextFactory)
+        public GetBukkenDetailsUseCase(IScrapingContextFactory scrapingContextFactory)
         {
             this._scrapingContextFactory = scrapingContextFactory;
         }
 
-        public async Task<IList<BukkenInfo>> ExecuteAsync(
-            FilterForm model,
+        public async Task<BukkenInfo> ExecuteAsync(
+            int id,
             CancellationToken cancellationToken = default
         )
         {
             using (var db = this._scrapingContextFactory.Create())
             {
-                return await db.GetFilteredBukkensAsync(model, cancellationToken)
-                    .ConfigureAwait(false);
+                return await db.GetBukkenDetailsAsync(id, cancellationToken).ConfigureAwait(false);
             }
         }
     }
