+++ SuumoScraping.Domain/Exceptions/SuumoExceptions.cs
