+++ SuumoScraping.Domain/Gateways/IScrapingContextFactory.cs
