+++ SuumoScraping.Domain/Gateways/ISuumoGateway.cs
