+++ SuumoScraping.Domain/Gateways/ISuumoHtmlFetcher.cs
