+++ SuumoScraping.Domain/Gateways/ISuumoHtmlParser.cs
