+++ SuumoScraping.Domain/Models/BukkenFile.cs
