+++ SuumoScraping.Domain/Models/Price.cs
@@ -4,20 +4,22 @@ namespace SuumoScraping.Domain.Models
     using System.Collections.Generic;
     using System.ComponentModel.DataAnnotations;
     using System.ComponentModel.DataAnnotations.Schema;
-    using Microsoft.EntityFrameworkCore;
 
-    [Table("newbukkenfiles")]
-    public partial class NewBukkenFile
+    [Table("prices")]
+    public partial class Price
     {
         [Key]
         public int Id { get; set; }
 
-        [Required]
-        [MaxLength(50)]
-        public string Type { get; set; }
+        public DateTime ChangedAt { get; set; }
 
-        [ForeignKey("File_Id")]
-        public virtual File File { get; set; }
+        [Column(TypeName = "decimal(18,2)")]
+        public decimal Min { get; set; }
+
+        [Column(TypeName = "decimal(18,2)")]
+        public decimal? Max { get; set; }
+
+        public string Text { get; set; }
 
         [ForeignKey("NewBukken_Id")]
         public virtual NewBukken NewBukken { get; set; }
