+++ SuumoScraping.Domain/Models/ScrapedBukken.cs
