+++ Views/Bukken/Details.cshtml
@@ -1,4 +1,4 @@
-@model SuumoScraping.ViewModels.BukkenInfo
+@model BukkenInfo
 
 @{
     ViewBag.Title = "Details";
