+++ Views/Bukken/Filter.cshtml
@@ -1,4 +1,4 @@
-@model SuumoScraping.ViewModels.FilterForm
+@model FilterForm
 
 @{
     ViewBag.Title = "検索条件";
