+++ Views/Bukken/List.cshtml
@@ -1,4 +1,4 @@
-﻿@model IReadOnlyList<SuumoScraping.ViewModels.BukkenInfo>
+@model IReadOnlyList<BukkenInfo>
 
 @{
     ViewBag.Title = "List";
