+++ Views/File/List.cshtml
@@ -1,4 +1,4 @@
-@model IReadOnlyList<SuumoScraping.ViewModels.FloorPlanInfo>
+@model IReadOnlyList<FloorPlanInfo>
 
 @{
     ViewBag.Title = "間取り一覧";
